+++ drake/examples/Cars/car_sim_lcm.cpp
@@ -48,7 +48,7 @@ int do_main(int argc, const char* argv[]) {
   auto weld_to_frame = allocate_shared<RigidBodyFrame>(
       aligned_allocator<RigidBodyFrame>(),
       // Weld the model to the world link.
-      "world",
+      RigidBodyTree::kWorldLinkName,
 
       // A pointer to a rigid body to which to weld the model is not needed
       // since the model will be welded to the world, which can by automatically
@@ -90,10 +90,10 @@ int do_main(int argc, const char* argv[]) {
 
   shared_ptr<lcm::LCM> lcm = make_shared<lcm::LCM>();
 
-  MatrixXd Kp(getNumInputs(*rigid_body_sys), tree->num_positions),
-      Kd(getNumInputs(*rigid_body_sys), tree->num_velocities);
+  MatrixXd Kp(getNumInputs(*rigid_body_sys), tree->number_of_positions()),
+      Kd(getNumInputs(*rigid_body_sys), tree->number_of_velocities());
   Matrix<double, Eigen::Dynamic, 3> map_driving_cmd_to_x_d(
-      tree->num_positions + tree->num_velocities, 3);
+      tree->number_of_positions() + tree->number_of_velocities(), 3);
   {  // setup PD controller for throttle and steering
     double kpSteering = 400, kdSteering = 80, kThrottle = 100;
     Kp.setZero();
@@ -115,9 +115,11 @@ int do_main(int argc, const char* argv[]) {
                  actuator_name == "left_wheel_joint") {
         auto const& b = tree->actuators[actuator_idx].body;
         Kd(actuator_idx, b->velocity_num_start) = kThrottle;  // throttle
-        map_driving_cmd_to_x_d(tree->num_positions + b->velocity_num_start, 1) =
+        map_driving_cmd_to_x_d(
+            tree->number_of_positions() + b->velocity_num_start, 1) =
             20;  // throttle (velocity) command
-        map_driving_cmd_to_x_d(tree->num_positions + b->velocity_num_start, 2) =
+        map_driving_cmd_to_x_d(
+            tree->number_of_positions() + b->velocity_num_start, 2) =
             -20;  // braking (velocity) command
       }
     }
@@ -142,7 +144,7 @@ int do_main(int argc, const char* argv[]) {
   options.timeout_seconds = numeric_limits<double>::infinity();
 
   VectorXd x0 = VectorXd::Zero(rigid_body_sys->getNumStates());
-  x0.head(tree->num_positions) = tree->getZeroConfiguration();
+  x0.head(tree->number_of_positions()) = tree->getZeroConfiguration();
   // todo:  call getInitialState instead?  (but currently, that would require
   // snopt).  needs #1627
   // I'm getting away without it, but might be generating large internal forces